% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqmc.R
\name{chisq.mc}
\alias{chisq.mc}
\title{Chi-sqaure post-hoc testing}
\usage{
chisq.mc(matrix, alpha = 0.05)
}
\arguments{
\item{mat}{A table or matrix}

\item{alpha}{Alpha significance value. Defaults to 0.05}
}
\description{
Perform Bonferroni-corrected chi-sqaure post-hoc testing on a table or matrix.
}
\examples{
chisq.mc(table(mtcars$gear, mtcars$cyl), alpha = 0.05)
}
