% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2num.R
\name{col2num}
\alias{col2num}
\title{Convert Excel column names to numbers}
\usage{
col2num(col)
}
\arguments{
\item{col}{A vector or an R object, which is coercible to one by as.vector(x, "character")}
}
\description{
When passed a letter, this function returns the associated column number. Used for calling Excel columns by name.
}
\examples{
col2num(c('T', 'AF'))
}
