% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_date.R
\name{excel_date}
\alias{excel_date}
\title{Transform dates imported from Excel}
\usage{
excel_date(42077, system = "windows")
}
\arguments{
\item{system}{Was data exported from Excel for Windows or Mac? Options: "windows", "mac"}
}
\description{
Creates a vector of randomly-generated, normally-distributed numbers with specified mean and sd.
}
\seealso{
as.Date
}
