% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFac.R
\name{mergeFac}
\alias{mergeFac}
\title{Merge and preserve level order of factors}
\usage{
mergeFac(...)
}
\arguments{
\item{...}{An array or an R object, which is coercible to one by as.array(x, "factor")}
}
\description{
concatenates factors while preserving original level order
}
