% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numfac.R
\name{numfac}
\alias{numfac}
\title{Convert factors to numeric}
\usage{
numfac(...)
}
\arguments{
\item{...}{An array or an R object, which is coercible to one by as.array(x, "factor")}
}
\description{
Converts factors to numeric. Replaces 'as.numeric(as.character(...))'
}
