% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Two-item permutations}
\usage{
permutation(x, repeats = FALSE)
}
\arguments{
\item{x}{A vector or an R object, which is coercible to one by as.vector(x, "numeric")}

\item{repeats}{logical. Can elements combine with themselves?}
}
\description{
Returns a vector of all possible 2-item permutations for elements in a vector
}
\examples{
permutation(x = c("a", "b", "c"), repeats = TRUE)
}
\seealso{
combn A function that executes this much better

expand.grid
}
