% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism.R
\name{prism}
\alias{prism}
\title{Export to GraphPad Prism and Excel}
\usage{
prism(formula, data, excel = F)
}
\arguments{
\item{formula}{A string}

\item{data}{A dataframe.}

\item{excel}{Pastes data into your clipboard for easy pasting into Excel. Defaults to False}
}
\description{
Perform Bonferroni-corrected chi-sqaure post-hoc testing on a table or matrix.
}
\examples{
prism("mpg ~ carb", mtcars, excel = T)
}

