% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NAs}
\usage{
replace_na(dat, repl)
}
\arguments{
\item{dat}{A data.frame or matrix}

\item{repl}{Replacement value. Defaults to 0.}
}
\description{
Replace all NA's with another value in a data.frame or matrix.
}
\examples{
replace_na(matrix(c(1,NA,3,4), ncol = 2), repl = 0)
}

