% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnorm.R
\name{vnorm}
\alias{vnorm}
\title{Normally-Distributed Vector}
\usage{
vnorm(100, 15, 3)
}
\arguments{
\item{n}{Number of items}

\item{mean}{Mean of the normal distribution}

\item{sd}{SD of the normal distribution}
}
\description{
Creates a vector of randomly-generated, normally-distributed numbers with specified mean and sd.
}
\examples{
x <- vnorm(100, 15, 3)
all(length(x) == 100, mean(x) == 15, sd(x) == 3)
}
\seealso{
scale

rnorm
}
