% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Find continuous sequences in discontinuous sequence}
\usage{
encode(x, print = FALSE)
}
\arguments{
\item{x}{A vector or an R object, which is coercible to one by as.vector(x, "numeric")}

\item{print}{logical. Should the list be printed to the console?}
}
\description{
When passed a vector of discontinuous numbers, this function returns a list of each continuous sequence
}
\examples{
encode(c(1:4, 7:10), print = TRUE)
}
